requirements.txt
fastpi
uvicorn
pydantic-settings
mysql-connector-python